- register in own activity - trying to resolve merge issue
