+++ android/app/src/main/java/de/fau/amos/virtualledger/android/App.java
@@ -28,7 +28,7 @@ public class App extends Application{
 
         netComponent = DaggerNetComponent.builder()
                 .appModule(new AppModule(this))
-                .netModule(new NetModule("http://192.168.0.56:8080"))
+                .netModule(new NetModule("http://192.168.178.21:8080"))
                 .build();
     }
